Выполнен lesson2/task1

+++ src/lesson1/task1/Simple.kt
@@ -59,7 +59,7 @@ fun main() {
  * Пользователь задает время в часах, минутах и секундах, например, 8:20:35.
  * Рассчитать время в секундах, прошедшее с начала суток (30035 в данном случае).
  */
-fun seconds(hours: Int, minutes: Int, seconds: Int): Int = 3600*hours + 60*minutes + seconds
+fun seconds(hours: Int, minutes: Int, seconds: Int): Int = 3600 * hours + 60 * minutes + seconds
 
 /**
  * Тривиальная
@@ -68,7 +68,7 @@ fun seconds(hours: Int, minutes: Int, seconds: Int): Int = 3600*hours + 60*minut
  * Определить длину того же отрезка в метрах (в данном случае 18.98).
  * 1 сажень = 3 аршина = 48 вершков, 1 вершок = 4.445 см.
  */
-fun lengthInMeters(sagenes: Int, arshins: Int, vershoks: Int): Double = 4.445*(sagenes*48 + arshins*48/3 + vershoks)/100
+fun lengthInMeters(sagenes: Int, arshins: Int, vershoks: Int): Double = 4.445 * (sagenes * 48 + arshins * 48 / 3 + vershoks) / 100
 
 /**
  * Тривиальная
@@ -76,7 +76,7 @@ fun lengthInMeters(sagenes: Int, arshins: Int, vershoks: Int): Double = 4.445*(s
  * Пользователь задает угол в градусах, минутах и секундах (например, 36 градусов 14 минут 35 секунд).
  * Вывести значение того же угла в радианах (например, 0.63256).
  */
-fun angleInRadian(deg: Int, min: Int, sec: Int): Double = (deg*3600+min*60+sec)* PI / 180 / 60 / 60
+fun angleInRadian(deg: Int, min: Int, sec: Int): Double = (deg * 3600 + min * 60 + sec) * PI / 180 / 60 / 60
 
 /**
  * Тривиальная
@@ -84,7 +84,7 @@ fun angleInRadian(deg: Int, min: Int, sec: Int): Double = (deg*3600+min*60+sec)*
  * Найти длину отрезка, соединяющего точки на плоскости с координатами (x1, y1) и (x2, y2).
  * Например, расстояние между (3, 0) и (0, 4) равно 5
  */
-fun trackLength(x1: Double, y1: Double, x2: Double, y2: Double): Double = sqrt(sqr(x1-x2)+sqr(y1-y2))
+fun trackLength(x1: Double, y1: Double, x2: Double, y2: Double): Double = sqrt(sqr(x1 - x2) + sqr(y1 - y2))
 
 /**
  * Простая
@@ -102,7 +102,7 @@ fun thirdDigit(number: Int): Int = number % 1000 / 100
  * Определите время поезда в пути в минутах (в данном случае 216).
  */
 fun travelMinutes(hoursDepart: Int, minutesDepart: Int, hoursArrive: Int, minutesArrive: Int): Int =
-    (hoursArrive-hoursDepart)*60 - minutesDepart + minutesArrive
+    (hoursArrive - hoursDepart) * 60 - minutesDepart + minutesArrive
 
 /**
  * Простая
@@ -111,7 +111,7 @@ fun travelMinutes(hoursDepart: Int, minutesDepart: Int, hoursArrive: Int, minute
  * Сколько денег будет на счету через 3 года (с учётом сложных процентов)?
  * Например, 100 рублей под 10% годовых превратятся в 133.1 рубля
  */
-fun accountInThreeYears(initial: Int, percent: Int): Double = (1.0+percent.toDouble()/100).pow(3)*initial
+fun accountInThreeYears(initial: Int, percent: Int): Double = (1.0 + percent / 100.0).pow(3) * initial
 
 /**
  * Простая
